Change bit_depth_image Asyn type to asynInt32
The $(P)$(R)BitDepthImage_RBV PV is a longin record for the
bit_depth_image SIMPLON API parameter, with DTYP=asynInt32, but the INP
parameter is EIG_DCD_bit_depth_image which indicates an Asyn type of
asynFloat64.  This causes the EigerParam::fetch(double &, int) member
function to be called to read the SIMPLON API parameter.  The member
function expects a JSON value_type of "float", but the bit_depth_image
parameter JSON value_type is either "int" or "uint", so the member
function fails with an error indicating that the JSON type is
unexpected.  Change the Asyn type of the bit_depth_image parameter to
asynInt32 to allow it to be read without error.